client/requirements.txt
cairocffi==1.5.1
CairoSVG==2.7.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
contourpy==1.0.7
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
fonttools==4.39.3
iconify==0.0.103
idna==2.10
kids.cache==0.0.7
kiwisolver==1.4.4
libsass==0.22.0
matplotlib==3.7.1
mock==5.0.2
numpy==1.24.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PySide2==5.15.2.1
python-dateutil==2.8.2
QT-PyQt-PySide-Custom-Widgets==0.6.4
QtPy==2.3.1
qtsass==0.4.0
requests==2.23.0
shiboken2==5.15.2.1
six==1.16.0
tinycss2==1.2.1
typing==3.7.4
urllib3==1.25.11
webencodings==0.5.1
